+++ examples/rediscli.py
@@ -1,5 +1,5 @@
 from flask import Flask
-from flask_exts import Manager
+from flask_exts import Exts
 from flask_exts.datastore.sqla import db
 from flask_exts.views.rediscli.view import RedisCli
 
@@ -9,14 +9,14 @@ from flask_exts.views.rediscli.mock_redis import MockRedis as Redis
 app = Flask(__name__)
 app.config["SECRET_KEY"] = "dev"
 
-# Manager init
-manager = Manager()
-manager.init_app(app)
+# Exts init
+exts = Exts()
+exts.init_app(app)
 
 # add rediscli view
 # redis_view = RedisCli(Redis())
 redis_view = RedisCli(Redis())
-manager.admin.add_view(redis_view)
+exts.admin.add_view(redis_view)
 
 with app.app_context():
     db.drop_all()