+++ src/flask_exts/__init__.py
@@ -1,1 +1,1 @@
-from .manager import Manager
+from .exts import Exts