+++ src/flask_exts/proxies.py
@@ -4,16 +4,16 @@ from flask import current_app
 from werkzeug.local import LocalProxy
 
 if t.TYPE_CHECKING:
-    from .manager import Manager
+    from .exts import Exts
     from .template.base import Template
     from .usercenter.base_user_store import BaseUserStore
     from .security.core import Security
 
 
-_manager: "Manager" = LocalProxy(lambda: current_app.extensions["manager"])
+_exts: "Exts" = LocalProxy(lambda: current_app.extensions["exts"])
 
-_template: "Template" = LocalProxy(lambda: _manager.template)
+_template: "Template" = LocalProxy(lambda: _exts.template)
 
-_userstore: "BaseUserStore" = LocalProxy(lambda: _manager.usercenter.userstore)
+_userstore: "BaseUserStore" = LocalProxy(lambda: _exts.usercenter.userstore)
 
-_security: "Security" = LocalProxy(lambda: _manager.security)
+_security: "Security" = LocalProxy(lambda: _exts.security)